+++ Dialogs/ExportSpectraDlg.cpp
@@ -10,7 +10,7 @@
 #include "ExportSpectraDlg.h"
 
 // Include the special multi-choice file-dialog
-#include "../Dialogs/FECFileDialog.h"
+#include "../Dialogs/CMultiSelectOpenFileDialog.h"
 
 using namespace Dialogs;
 using namespace novac;
