+++ Dialogs/ImportSpectraDlg.cpp
@@ -9,7 +9,7 @@
 #include <CDerr.h>
 
 // Include the special multi-choice file-dialog
-#include "../Dialogs/FECFileDialog.h"
+#include "../Dialogs/CMultiSelectOpenFileDialog.h"
 
 // CImportSpectraDlg dialog
