+++ Dialogs/MergePakFilesDlg.cpp
@@ -8,7 +8,7 @@
 #include "../Common/Spectra/PakFileHandler.h"
 
 // Include the special multi-choice file-dialog
-#include "../Dialogs/FECFileDialog.h"
+#include "../Dialogs/CMultiSelectOpenFileDialog.h"
 
 using namespace Dialogs;
 using namespace novac;
