+++ Dialogs/SplitPakFilesDlg.cpp
@@ -9,7 +9,7 @@
 #include "../Common/Spectra/PakFileHandler.h"
 
 // Include the special multi-choice file-dialog
-#include "../Dialogs/FECFileDialog.h"
+#include "../Dialogs/CMultiSelectOpenFileDialog.h"
 
 using namespace Dialogs;
 // CSplitPakFilesDlg dialog
